working on compile.cpp
adding folding for all of the combinations of the builtin integer
types.

+++ inf/include/imr/ssa.hpp
@@ -21,7 +21,7 @@
 
 namespace inf {
 struct ssa {
-    uint32_t index;
+    uint64_t index;
 };
 } // namespace inf
